Blues service discovery : Search device in dbus before deciding to create it
Previously, in blues, we tried to create the dbus device and if it already
existed then we searched / used it. Now the behaviour is reversed.

It's better to do it this way because sometimes a non privileged user
cannot create the device (org.freedesktop.DBus.Error.AuthFailed) as it is
the case in SailfishOS.